backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23

# Face Recognition
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
dlib==19.22.1
numpy==1.24.3
scipy==1.10.1

# Data Processing
pandas==2.0.3
openpyxl==3.1.2

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Utils
python-dateutil==2.8.2

# Security - JWT认证与密码哈希
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Security - 速率限制与安全头
slowapi==0.1.9
starlette-secure-headers==0.3.0

# ===== 运维增强 =====
# 结构化日志
structlog==24.4.0

# 数据库迁移
alembic==1.13.1

# Prometheus 监控
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.19.0

# ===== 性能优化 =====
# Redis 缓存（性能优化 Phase 3）
redis[hiredis]==5.0.4
